ReverseSlide function added, small Dots bug fixed
ReverseSlide: Allows the slider to be tethered on the right and dragged
from the left.
Dots bug: When `step=null` while `dots=true` the Dots code breaks your
browser by trying to add null. Also, dots are added even when
`dots=false` if you pass marks as a param to the slider. I have fixed
both of these issues.