Allow callback for "dotStyle" and "activeDotStyle" props
I have to use rc-slider to make a range selector, where the active range has a horizontal color gradient. I can style the trackline with pure CSS, but not the step dots. I even went to great extents with CSS by adding some creative attribute selectors based on the `style="left: ..."` values, which was enough to have the correct dot colors when the full range is selected. But when the range is set to a narrower state, the gradient changes too, and the dots can't adapt to that. I can only solve this with a callback where I can set the style based on the dot position and the slider value. This is what this PR is about.